style: 하단 네비게이션바 수정 (menuName 추가)
