+++ apps/client/src/components/layouts/GnbNavBar.tsx
@@ -9,10 +9,11 @@ export default function GnbNavBar() {
     <div>
       <div className="fixed bottom-0 rounded-t-5xl max-w-[600px] w-full h-[80px] bg-gradient-to-t from-primary" />
       <nav className="fixed bottom-0 rounded-t-3xl max-w-[600px] w-full h-[60px] bg-white">
-        <ul className="relative flex items-center justify-evenly px-2 h-full">
+        <ul className="relative flex items-center justify-between h-full px-5">
           {gnbMenuData.map((menu) => (
             <GnbMenu
-              key={menu.link}
+              key={menu.menuName}
+              menuName={menu.menuName}
               link={menu.link}
               icon={menu.icon}
               main={menu.main || false}
