--- apps/client/src/types/gnbMenuType.ts
@@ -1,5 +1,0 @@
-export interface GnbMenuType {
-  main?: boolean;
-  link: string;
-  icon: React.FC<{ className?: string; onClick?: () => void }>;
-}
