--- apps/client/src/types/homeTabType.ts
@@ -1,12 +1,0 @@
-export interface HomeTabType {
-  tabMenuName: string;
-  onClick?: () => void;
-  selected?: boolean;
-}
-
-export interface HomeTabBarProps {
-  tabContents: {
-    reservationParking: React.ReactNode;
-    currentParking: React.ReactNode;
-  };
-}
