--- apps/client/src/types/markerDataType.ts
@@ -1,6 +1,0 @@
-export interface markerDataType {
-  parkingLotUuid: string;
-  availableSpots: number;
-  latitude: number;
-  longitude: number;
-}
